+++ imagepicker/src/main/kotlin/com/github/drjacky/imagepicker/provider/CameraProvider.kt
@@ -5,8 +5,8 @@ import android.app.Activity
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
+import androidx.activity.result.ActivityResult
 import androidx.core.app.ActivityCompat.requestPermissions
-import com.github.drjacky.imagepicker.ImagePicker
 import com.github.drjacky.imagepicker.ImagePickerActivity
 import com.github.drjacky.imagepicker.R
 import com.github.drjacky.imagepicker.util.FileUtil
@@ -22,7 +22,11 @@ import java.io.File
  * @version 1.0
  * @since 04 January 2019
  */
-class CameraProvider(activity: ImagePickerActivity, private val tryFrontCamera: Boolean = false) :
+class CameraProvider(
+    activity: ImagePickerActivity,
+    private val tryFrontCamera: Boolean = false,
+    private val launcher: (Intent) -> Unit
+) :
     BaseProvider(activity) {
 
     companion object {
@@ -46,7 +50,6 @@ class CameraProvider(activity: ImagePickerActivity, private val tryFrontCamera: 
             Manifest.permission.CAMERA
         )
 
-        private const val CAMERA_INTENT_REQ_CODE = 4281
         private const val PERMISSION_INTENT_REQ_CODE = 4282
     }
 
@@ -60,21 +63,6 @@ class CameraProvider(activity: ImagePickerActivity, private val tryFrontCamera: 
      */
     private val mAskCameraPermission = PermissionUtil
         .isPermissionInManifest(this, Manifest.permission.CAMERA)
-
-    /**
-     * Camera image will be stored in below file directory
-     */
-    private var mFileDir: File? = null
-
-    init {
-        val bundle = activity.intent.extras ?: Bundle()
-
-        // Get File Directory
-        val fileDir = bundle.getString(ImagePicker.EXTRA_SAVE_DIRECTORY)
-        fileDir?.let {
-            mFileDir = File(it)
-        }
-    }
 
     /**
      * Save CameraProvider state
@@ -133,13 +121,12 @@ class CameraProvider(activity: ImagePickerActivity, private val tryFrontCamera: 
      */
     private fun startCameraIntent() {
         // Create and get empty file to store capture image content
-        val file = FileUtil.getImageFile(dir = mFileDir)
+        val file = FileUtil.getImageFile(this)
         mCameraFile = file
 
         // Check if file exists
         if (file != null && file.exists()) {
-            val cameraIntent = IntentUtils.getCameraIntent(this, file, tryFrontCamera)
-            activity.startActivityForResult(cameraIntent, CAMERA_INTENT_REQ_CODE)
+            launcher.invoke(IntentUtils.getCameraIntent(this, file, tryFrontCamera))
         } else {
             setError(R.string.error_failed_to_create_camera_image_file)
         }
@@ -166,28 +153,12 @@ class CameraProvider(activity: ImagePickerActivity, private val tryFrontCamera: 
         }
     }
 
-    /**
-     * Handle Camera Intent Activity Result
-     *
-     * @param requestCode It must be {@link CameraProvider#CAMERA_INTENT_REQ_CODE}
-     * @param resultCode For success it should be {@link Activity#RESULT_OK}
-     * @param data Result Intent
-     */
-    fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        if (requestCode == CAMERA_INTENT_REQ_CODE) {
-            if (resultCode == Activity.RESULT_OK) {
-                handleResult(data)
-            } else {
-                setResultCancel()
-            }
+    fun handleResult(result: ActivityResult) {
+        if (result.resultCode == Activity.RESULT_OK) {
+            activity.setImage(mCameraFile!!)
+        } else {
+            setResultCancel()
         }
-    }
-
-    /**
-     * This method will be called when final result fot this provider is enabled.
-     */
-    private fun handleResult(data: Intent?) {
-        activity.setImage(mCameraFile!!)
     }
 
     /**