+++ imagepicker/src/main/kotlin/com/github/drjacky/imagepicker/provider/CropProvider.kt
@@ -1,11 +1,12 @@
 package com.github.drjacky.imagepicker.provider
 
 import android.app.Activity
-import android.content.ActivityNotFoundException
+import android.content.Context
 import android.content.Intent
 import android.net.Uri
 import android.os.Bundle
 import android.util.Log
+import androidx.activity.result.ActivityResult
 import com.github.drjacky.imagepicker.ImagePicker
 import com.github.drjacky.imagepicker.ImagePickerActivity
 import com.github.drjacky.imagepicker.R
@@ -22,44 +23,31 @@ import java.io.IOException
  * @version 1.0
  * @since 04 January 2019
  */
-class CropProvider(activity: ImagePickerActivity) : BaseProvider(activity) {
+class CropProvider(activity: ImagePickerActivity, private val launcher: (Intent) -> Unit) :
+    BaseProvider(activity) {
 
     companion object {
         private val TAG = CropProvider::class.java.simpleName
-
-        /**
-         * Key to Save/Retrieve Crop File state
-         */
         private const val STATE_CROP_FILE = "state.crop_file"
     }
 
-    private val mMaxWidth: Int
-    private val mMaxHeight: Int
+    private val maxWidth: Int
+    private val maxHeight: Int
 
-    private val mCropOval: Boolean
-    private val mCrop: Boolean
-    private val mCropAspectX: Float
-    private val mCropAspectY: Float
-    private var mCropImageFile: File? = null
-    private var mFileDir: File? = null
+    private val cropOval: Boolean
+    private val crop: Boolean
+    private val cropAspectX: Float
+    private val cropAspectY: Float
+    private var cropImageFile: File? = null
 
     init {
-        val bundle = activity.intent.extras ?: Bundle()
-
-        // Get Max Width/Height parameter from Intent
-        mMaxWidth = bundle.getInt(ImagePicker.EXTRA_MAX_WIDTH, 0)
-        mMaxHeight = bundle.getInt(ImagePicker.EXTRA_MAX_HEIGHT, 0)
-
-        // Get Crop Aspect Ratio parameter from Intent
-        mCropOval = bundle.getBoolean(ImagePicker.EXTRA_CROP_OVAL, false)
-        mCrop = bundle.getBoolean(ImagePicker.EXTRA_CROP, false)
-        mCropAspectX = bundle.getFloat(ImagePicker.EXTRA_CROP_X, 0f)
-        mCropAspectY = bundle.getFloat(ImagePicker.EXTRA_CROP_Y, 0f)
-
-        // Get File Directory
-        val fileDir = bundle.getString(ImagePicker.EXTRA_SAVE_DIRECTORY)
-        fileDir?.let {
-            mFileDir = File(it)
+        with(activity.intent.extras ?: Bundle()) {
+            maxWidth = getInt(ImagePicker.EXTRA_MAX_WIDTH, 0)
+            maxHeight = getInt(ImagePicker.EXTRA_MAX_HEIGHT, 0)
+            crop = getBoolean(ImagePicker.EXTRA_CROP, false)
+            cropOval = getBoolean(ImagePicker.EXTRA_CROP_OVAL, false)
+            cropAspectX = getFloat(ImagePicker.EXTRA_CROP_X, 0f)
+            cropAspectY = getFloat(ImagePicker.EXTRA_CROP_Y, 0f)
         }
     }
 
@@ -74,16 +62,14 @@ class CropProvider(activity: ImagePickerActivity) : BaseProvider(activity) {
      * Note: To produce this scenario, enable "Don't keep activities" from developer options
      */
     override fun onSaveInstanceState(outState: Bundle) {
-        // Save crop file
-        outState.putSerializable(STATE_CROP_FILE, mCropImageFile)
+        outState.putSerializable(STATE_CROP_FILE, cropImageFile)
     }
 
     /**
      * Retrieve CropProvider state
      */
     override fun onRestoreInstanceState(savedInstanceState: Bundle?) {
-        // Restore crop file
-        mCropImageFile = savedInstanceState?.getSerializable(STATE_CROP_FILE) as File?
+        cropImageFile = savedInstanceState?.getSerializable(STATE_CROP_FILE) as File?
     }
 
     /**
@@ -91,20 +77,20 @@ class CropProvider(activity: ImagePickerActivity) : BaseProvider(activity) {
      *
      * @return Boolean. True if it is allow dimmed layer to have a circle inside else false.
      */
-    fun isCropOvalEnabled() = mCropOval
+    fun isCropOvalEnabled() = cropOval
 
     /**
      * Check if crop should be enabled or not
      *
      * @return Boolean. True if Crop should be enabled else false.
      */
-    fun isCropEnabled() = mCrop
+    fun isCropEnabled() = crop
 
     /**
      * Start Crop Activity
      */
-    fun startIntent(file: File, cropOval: Boolean) {
-        cropImage(file, cropOval)
+    fun startIntent(context: Context, file: File, cropOval: Boolean) {
+        cropImage(context, file, cropOval)
     }
 
     /**
@@ -112,12 +98,12 @@ class CropProvider(activity: ImagePickerActivity) : BaseProvider(activity) {
      * @throws IOException if failed to crop image
      */
     @Throws(IOException::class)
-    private fun cropImage(file: File, cropOval: Boolean) {
+    private fun cropImage(context: Context, file: File, cropOval: Boolean) {
         val uri = Uri.fromFile(file)
         val extension = FileUriUtils.getImageExtension(uri)
-        mCropImageFile = FileUtil.getImageFile(dir = mFileDir, extension = extension)
+        cropImageFile = FileUtil.getImageFile(context = context, extension = extension)
 
-        if (mCropImageFile == null || !mCropImageFile!!.exists()) {
+        if (cropImageFile == null || !cropImageFile!!.exists()) {
             Log.e(TAG, "Failed to create crop image file")
             setError(R.string.error_failed_to_crop_image)
             return
@@ -126,59 +112,32 @@ class CropProvider(activity: ImagePickerActivity) : BaseProvider(activity) {
         val options = UCrop.Options()
         options.setCompressionFormat(FileUtil.getCompressFormat(extension))
         options.setCircleDimmedLayer(cropOval)
-        val uCrop = UCrop.of(uri, Uri.fromFile(mCropImageFile))
-            .withOptions(options)
+        val uCrop = UCrop.of(uri, Uri.fromFile(cropImageFile)).withOptions(options)
 
-        if (mCropAspectX > 0 && mCropAspectY > 0) {
-            uCrop.withAspectRatio(mCropAspectX, mCropAspectY)
+        if (cropAspectX > 0 && cropAspectY > 0) {
+            uCrop.withAspectRatio(cropAspectX, cropAspectY)
         }
 
-        if (mMaxWidth > 0 && mMaxHeight > 0) {
-            uCrop.withMaxResultSize(mMaxWidth, mMaxHeight)
+        if (maxWidth > 0 && maxHeight > 0) {
+            uCrop.withMaxResultSize(maxWidth, maxHeight)
         }
 
-        try {
-            uCrop.start(activity, UCrop.REQUEST_CROP)
-        } catch (ex: ActivityNotFoundException) {
-            setError(
-                "uCrop not specified in manifest file." +
-                        "Add UCropActivity in Manifest" +
-                        "<activity\n" +
-                        "    android:name=\"com.yalantis.ucrop.UCropActivity\"\n" +
-                        "    android:screenOrientation=\"portrait\"\n" +
-                        "    android:theme=\"@style/Theme.AppCompat.Light.NoActionBar\"/>"
-            )
-            ex.printStackTrace()
-        }
-    }
-
-    /**
-     * Handle Crop Intent Activity Result
-     *
-     * @param requestCode It must be {@link UCrop#REQUEST_CROP}
-     * @param resultCode For success it should be {@link Activity#RESULT_OK}
-     * @param data Result Intent
-     */
-    fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        if (requestCode == UCrop.REQUEST_CROP) {
-            if (resultCode == Activity.RESULT_OK) {
-                handleResult(mCropImageFile)
-            } else {
-                setResultCancel()
-            }
-        }
+        launcher.invoke(uCrop.getIntent(activity))
     }
 
     /**
      * This method will be called when final result fot this provider is enabled.
-     *
-     * @param file cropped file
      */
-    private fun handleResult(file: File?) {
-        if (file != null) {
-            activity.setCropImage(file)
+    fun handleResult(result: ActivityResult) {
+        if (result.resultCode == Activity.RESULT_OK) {
+            val file = cropImageFile
+            if (file != null) {
+                activity.setCropImage(file)
+            } else {
+                setError(R.string.error_failed_to_crop_image)
+            }
         } else {
-            setError(R.string.error_failed_to_crop_image)
+            setResultCancel()
         }
     }
 
@@ -186,6 +145,6 @@ class CropProvider(activity: ImagePickerActivity) : BaseProvider(activity) {
      * Delete Crop file is exists
      */
     override fun onFailure() {
-        mCropImageFile?.delete()
+        cropImageFile?.delete()
     }
 }