+++ imagepicker/src/main/kotlin/com/github/drjacky/imagepicker/util/FileUtil.kt
@@ -1,5 +1,6 @@
 package com.github.drjacky.imagepicker.util
 
+import android.content.Context
 import android.graphics.Bitmap
 import android.os.Environment
 import android.os.StatFs
@@ -27,14 +28,14 @@ object FileUtil {
      * @return Return Empty file to store camera image.
      * @throws IOException if permission denied of failed to create new file.
      */
-    fun getImageFile(dir: File? = null, extension: String? = null): File? {
+    fun getImageFile(context: Context, dir: File? = null, extension: String? = null): File? {
         try {
             // Create an image file name
             val ext = extension ?: ".jpg"
             val imageFileName = "IMG_${getTimestamp()}$ext"
 
             // Create File Directory Object
-            val storageDir = dir ?: getCameraDirectory()
+            val storageDir = dir ?: getCameraDirectory(context)
 
             // Create Directory If not exist
             if (!storageDir.exists()) storageDir.mkdirs()
@@ -57,8 +58,9 @@ object FileUtil {
      *
      * @return File Camera Image Directory
      */
-    private fun getCameraDirectory(): File {
-        val dir = Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DCIM)
+    private fun getCameraDirectory(context: Context): File {
+        val dir =
+            context.getExternalFilesDir(Environment.DIRECTORY_DCIM)// Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DCIM)
         return File(dir, "Camera")
     }