+++ sample/src/main/kotlin/com.github.drjacky.imagepicker/sample/MainActivity.kt
@@ -1,14 +1,13 @@
 package com.github.drjacky.imagepicker.sample
 
 import android.app.Activity
-import android.content.Intent
 import android.os.Bundle
-import android.os.Environment
-import android.util.Log
 import android.view.Menu
 import android.view.MenuItem
 import android.view.View
 import android.widget.Toast
+import androidx.activity.result.ActivityResult
+import androidx.activity.result.contract.ActivityResultContracts
 import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import com.github.drjacky.imagepicker.ImagePicker
@@ -23,17 +22,47 @@ import java.io.File
 class MainActivity : AppCompatActivity() {
 
     companion object {
-
         private const val GITHUB_REPOSITORY = "https://github.com/drjacky/ImagePicker"
-
-        private const val PROFILE_IMAGE_REQ_CODE = 101
-        private const val GALLERY_IMAGE_REQ_CODE = 102
-        private const val CAMERA_IMAGE_REQ_CODE = 103
     }
 
     private var mCameraFile: File? = null
     private var mGalleryFile: File? = null
     private var mProfileFile: File? = null
+
+    private val profileLauncher =
+        registerForActivityResult(ActivityResultContracts.StartActivityForResult()) {
+            if (it.resultCode == Activity.RESULT_OK) {
+                val file = ImagePicker.getFile(it.data)!!
+                mProfileFile = file
+                imgProfile.setLocalImage(file, true)
+            } else parseError(it)
+        }
+    private val galleryLauncher =
+        registerForActivityResult(ActivityResultContracts.StartActivityForResult()) {
+            if (it.resultCode == Activity.RESULT_OK) {
+                val file = ImagePicker.getFile(it.data)!!
+                mGalleryFile = file
+                imgGallery.setLocalImage(file)
+            } else parseError(it)
+        }
+
+    private val cameraLauncher =
+        registerForActivityResult(ActivityResultContracts.StartActivityForResult()) {
+            if (it.resultCode == Activity.RESULT_OK) {
+                val file = ImagePicker.getFile(it.data)!!
+                mCameraFile = file
+                imgCamera.setLocalImage(file, false)
+            } else parseError(it)
+        }
+
+    private fun parseError(activityResult: ActivityResult) {
+        if (activityResult.resultCode == ImagePicker.RESULT_ERROR) {
+            Toast.makeText(this, ImagePicker.getError(activityResult.data), Toast.LENGTH_SHORT)
+                .show()
+        } else {
+            Toast.makeText(this, "Task Cancelled", Toast.LENGTH_SHORT).show()
+        }
+    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -59,74 +88,35 @@ class MainActivity : AppCompatActivity() {
 
     fun pickProfileImage(view: View) {
         ImagePicker.with(this)
-            // Crop Square image
-            .cropSquare()
-            // Allow dimmed layer to have a circle inside
+            .crop()
             .cropOval()
-            .setImageProviderInterceptor { imageProvider -> // Intercept ImageProvider
-                Log.d("ImagePicker", "Selected ImageProvider: " + imageProvider.name)
-            }
-            // Image resolution will be less than 512 x 512
-            .maxResultSize(512, 512)
-            .start(PROFILE_IMAGE_REQ_CODE)
+            .maxResultSize(512, 512, true)
+            .createIntentFromDialog { profileLauncher.launch(it) }
     }
 
     fun pickGalleryImage(view: View) {
-        ImagePicker.with(this)
-            // Crop Image(User can choose Aspect Ratio)
-            .crop()
-            // User can only select image from Gallery
-            .galleryOnly()
-
-            .galleryMimeTypes(  //no gif images at all
-                mimeTypes = arrayOf(
-                    "image/png",
-                    "image/jpg",
-                    "image/jpeg"
+        galleryLauncher.launch(
+            ImagePicker.with(this)
+                //.crop()
+                .galleryOnly()
+                .galleryMimeTypes(  //no gif images at all
+                    mimeTypes = arrayOf(
+                        "image/png",
+                        "image/jpg",
+                        "image/jpeg"
+                    )
                 )
-            )
-            // Image resolution will be less than 1080 x 1920
-            .maxResultSize(1080, 1920)
-            .start(GALLERY_IMAGE_REQ_CODE)
+                .createIntent()
+        )
     }
 
     fun pickCameraImage(view: View) {
-        ImagePicker.with(this)
-            // User can only capture image from Camera
-            .cameraOnly()
-            // Image size will be less than 1024 KB
-            .compress(1024)
-            .saveDir(Environment.getExternalStorageDirectory())
-            // .saveDir(Environment.getExternalStorageDirectory().absolutePath+File.separator+"ImagePicker")
-            // .saveDir(getExternalFilesDir(null)!!)
-            .start(CAMERA_IMAGE_REQ_CODE)
-    }
-
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        super.onActivityResult(requestCode, resultCode, data)
-        if (resultCode == Activity.RESULT_OK) {
-            Log.e("TAG", "Path:${ImagePicker.getFilePath(data)}")
-            // File object will not be null for RESULT_OK
-            val file = ImagePicker.getFile(data)!!
-            when (requestCode) {
-                PROFILE_IMAGE_REQ_CODE -> {
-                    mProfileFile = file
-                    imgProfile.setLocalImage(file, true)
-                }
-                GALLERY_IMAGE_REQ_CODE -> {
-                    mGalleryFile = file
-                    imgGallery.setLocalImage(file)
-                }
-                CAMERA_IMAGE_REQ_CODE -> {
-                    mCameraFile = file
-                    imgCamera.setLocalImage(file, false)
-                }
-            }
-        } else if (resultCode == ImagePicker.RESULT_ERROR) {
-            Toast.makeText(this, ImagePicker.getError(data), Toast.LENGTH_SHORT).show()
-        } else {
-            Toast.makeText(this, "Task Cancelled", Toast.LENGTH_SHORT).show()
-        }
+        cameraLauncher.launch(
+            ImagePicker.with(this)
+                .cameraOnly()
+                .maxResultSize(1080, 1920, true)
+                .createIntent()
+        )
     }
 
     fun showImageCode(view: View) {